connector data coll refactor
A quick code movement before adding the
sharepoint datacollection production so that
we minimize graph_conn file bloat.
