+++ src/internal/connector/graph_connector_test.go
@@ -1,8 +1,6 @@
 package connector
 
 import (
-	"bytes"
-	"context"
 	"testing"
 	"time"
 
@@ -12,7 +10,6 @@ import (
 
 	"github.com/alcionai/corso/src/internal/connector/exchange"
 	"github.com/alcionai/corso/src/internal/connector/mockconnector"
-	"github.com/alcionai/corso/src/internal/connector/support"
 	"github.com/alcionai/corso/src/internal/data"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/pkg/control"
@@ -24,14 +21,6 @@ type GraphConnectorIntegrationSuite struct {
 	suite.Suite
 	connector *GraphConnector
 	user      string
-}
-
-func loadConnector(ctx context.Context, t *testing.T) *GraphConnector {
-	a := tester.NewM365Account(t)
-	connector, err := NewGraphConnector(ctx, a)
-	require.NoError(t, err)
-
-	return connector
 }
 
 func TestGraphConnectorIntegrationSuite(t *testing.T) {
@@ -78,287 +67,6 @@ func (suite *GraphConnectorIntegrationSuite) TestSetTenantUsers() {
 	err = newConnector.setTenantUsers(ctx)
 	assert.NoError(suite.T(), err)
 	suite.Greater(len(newConnector.Users), 0)
-}
-
-// TestInvalidUserForDataCollections ensures verification process for users
-func (suite *GraphConnectorIntegrationSuite) TestInvalidUserForDataCollections() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	invalidUser := "foo@example.com"
-	connector := loadConnector(ctx, suite.T())
-	tests := []struct {
-		name        string
-		getSelector func(t *testing.T) selectors.Selector
-	}{
-		{
-			name: "invalid exchange backup user",
-			getSelector: func(t *testing.T) selectors.Selector {
-				sel := selectors.NewExchangeBackup()
-				sel.Include(sel.MailFolders([]string{invalidUser}, selectors.Any()))
-				return sel.Selector
-			},
-		},
-		{
-			name: "Invalid onedrive backup user",
-			getSelector: func(t *testing.T) selectors.Selector {
-				sel := selectors.NewOneDriveBackup()
-				sel.Include(sel.Folders([]string{invalidUser}, selectors.Any()))
-				return sel.Selector
-			},
-		},
-	}
-
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			collections, err := connector.DataCollections(ctx, test.getSelector(t))
-			assert.Error(t, err)
-			assert.Empty(t, collections)
-		})
-	}
-}
-
-// TestExchangeDataCollection verifies interface between operation and
-// GraphConnector remains stable to receive a non-zero amount of Collections
-// for the Exchange Package. Enabled exchange applications:
-// - mail
-// - contacts
-// - events
-func (suite *GraphConnectorIntegrationSuite) TestExchangeDataCollection() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	connector := loadConnector(ctx, suite.T())
-	tests := []struct {
-		name        string
-		getSelector func(t *testing.T) selectors.Selector
-	}{
-		{
-			name: suite.user + " Email",
-			getSelector: func(t *testing.T) selectors.Selector {
-				sel := selectors.NewExchangeBackup()
-				sel.Include(sel.MailFolders([]string{suite.user}, []string{exchange.DefaultMailFolder}, selectors.PrefixMatch()))
-
-				return sel.Selector
-			},
-		},
-		{
-			name: suite.user + " Contacts",
-			getSelector: func(t *testing.T) selectors.Selector {
-				sel := selectors.NewExchangeBackup()
-				sel.Include(sel.ContactFolders(
-					[]string{suite.user},
-					[]string{exchange.DefaultContactFolder},
-					selectors.PrefixMatch()))
-
-				return sel.Selector
-			},
-		},
-		// {
-		// 	name: suite.user + " Events",
-		// 	getSelector: func(t *testing.T) selectors.Selector {
-		// 		sel := selectors.NewExchangeBackup()
-		// 		sel.Include(sel.EventCalendars([]string{suite.user},
-		//		[]string{exchange.DefaultCalendar}, selectors.PrefixMatch()))
-
-		// 		return sel.Selector
-		// 	},
-		// },
-	}
-
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			collection, err := connector.ExchangeDataCollection(ctx, test.getSelector(t))
-			require.NoError(t, err)
-			assert.Equal(t, len(collection), 1)
-			channel := collection[0].Items()
-			for object := range channel {
-				buf := &bytes.Buffer{}
-				_, err := buf.ReadFrom(object.ToReader())
-				assert.NoError(t, err, "received a buf.Read error")
-			}
-			status := connector.AwaitStatus()
-			assert.NotZero(t, status.Successful)
-			t.Log(status.String())
-		})
-	}
-}
-
-// TestMailSerializationRegression verifies that all mail data stored in the
-// test account can be successfully downloaded into bytes and restored into
-// M365 mail objects
-func (suite *GraphConnectorIntegrationSuite) TestMailSerializationRegression() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	t := suite.T()
-	connector := loadConnector(ctx, t)
-	sel := selectors.NewExchangeBackup()
-	sel.Include(sel.MailFolders([]string{suite.user}, []string{exchange.DefaultMailFolder}, selectors.PrefixMatch()))
-	collection, err := connector.createCollections(ctx, sel.Scopes()[0])
-	require.NoError(t, err)
-
-	for _, edc := range collection {
-		suite.T().Run(edc.FullPath().String(), func(t *testing.T) {
-			streamChannel := edc.Items()
-			// Verify that each message can be restored
-			for stream := range streamChannel {
-				buf := &bytes.Buffer{}
-				read, err := buf.ReadFrom(stream.ToReader())
-				assert.NoError(t, err)
-				assert.NotZero(t, read)
-				message, err := support.CreateMessageFromBytes(buf.Bytes())
-				assert.NotNil(t, message)
-				assert.NoError(t, err)
-			}
-		})
-	}
-
-	status := connector.AwaitStatus()
-	suite.NotNil(status)
-	suite.Equal(status.ObjectCount, status.Successful)
-}
-
-// TestContactSerializationRegression verifies ability to query contact items
-// and to store contact within Collection. Downloaded contacts are run through
-// a regression test to ensure that downloaded items can be uploaded.
-func (suite *GraphConnectorIntegrationSuite) TestContactSerializationRegression() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	connector := loadConnector(ctx, suite.T())
-
-	tests := []struct {
-		name          string
-		getCollection func(t *testing.T) []*exchange.Collection
-	}{
-		{
-			name: "Default Contact Folder",
-			getCollection: func(t *testing.T) []*exchange.Collection {
-				scope := selectors.
-					NewExchangeBackup().
-					ContactFolders([]string{suite.user}, []string{exchange.DefaultContactFolder}, selectors.PrefixMatch())[0]
-				collections, err := connector.createCollections(ctx, scope)
-				require.NoError(t, err)
-
-				return collections
-			},
-		},
-	}
-
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			edcs := test.getCollection(t)
-			require.Equal(t, len(edcs), 1)
-			edc := edcs[0]
-			assert.Equal(t, edc.FullPath().Folder(), exchange.DefaultContactFolder)
-			streamChannel := edc.Items()
-			count := 0
-			for stream := range streamChannel {
-				buf := &bytes.Buffer{}
-				read, err := buf.ReadFrom(stream.ToReader())
-				assert.NoError(t, err)
-				assert.NotZero(t, read)
-				contact, err := support.CreateContactFromBytes(buf.Bytes())
-				assert.NotNil(t, contact)
-				assert.NoError(t, err, "error on converting contact bytes: "+string(buf.Bytes()))
-				count++
-			}
-			assert.NotZero(t, count)
-
-			status := connector.AwaitStatus()
-			suite.NotNil(status)
-			suite.Equal(status.ObjectCount, status.Successful)
-		})
-	}
-}
-
-// TestEventsSerializationRegression ensures functionality of createCollections
-// to be able to successfully query, download and restore event objects
-func (suite *GraphConnectorIntegrationSuite) TestEventsSerializationRegression() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	connector := loadConnector(ctx, suite.T())
-
-	tests := []struct {
-		name, expected string
-		getCollection  func(t *testing.T) []*exchange.Collection
-	}{
-		{
-			name:     "Default Event Calendar",
-			expected: exchange.DefaultCalendar,
-			getCollection: func(t *testing.T) []*exchange.Collection {
-				sel := selectors.NewExchangeBackup()
-				sel.Include(sel.EventCalendars([]string{suite.user}, []string{exchange.DefaultCalendar}, selectors.PrefixMatch()))
-				collections, err := connector.createCollections(ctx, sel.Scopes()[0])
-				require.NoError(t, err)
-
-				return collections
-			},
-		},
-		{
-			name:     "Birthday Calendar",
-			expected: "Birthdays",
-			getCollection: func(t *testing.T) []*exchange.Collection {
-				sel := selectors.NewExchangeBackup()
-				sel.Include(sel.EventCalendars([]string{suite.user}, []string{"Birthdays"}, selectors.PrefixMatch()))
-				collections, err := connector.createCollections(ctx, sel.Scopes()[0])
-				require.NoError(t, err)
-
-				return collections
-			},
-		},
-	}
-
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			collections := test.getCollection(t)
-			require.Equal(t, len(collections), 1)
-			edc := collections[0]
-			assert.Equal(t, edc.FullPath().Folder(), test.expected)
-			streamChannel := edc.Items()
-
-			for stream := range streamChannel {
-				buf := &bytes.Buffer{}
-				read, err := buf.ReadFrom(stream.ToReader())
-				assert.NoError(t, err)
-				assert.NotZero(t, read)
-				event, err := support.CreateEventFromBytes(buf.Bytes())
-				assert.NotNil(t, event)
-				assert.NoError(t, err, "experienced error parsing event bytes: "+string(buf.Bytes()))
-			}
-
-			status := connector.AwaitStatus()
-			suite.NotNil(status)
-			suite.Equal(status.ObjectCount, status.Successful)
-		})
-	}
-}
-
-// TestAccessOfInboxAllUsers verifies that GraphConnector can
-// support `--users *` for backup operations. Selector.DiscreteScopes
-// returns all of the users within one scope. Only users who have
-// messages in their inbox will have a collection returned.
-// The final test insures that more than a 75% of the user collections are
-// returned. If an error was experienced, the test will fail overall
-func (suite *GraphConnectorIntegrationSuite) TestAccessOfInboxAllUsers() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	t := suite.T()
-	connector := loadConnector(ctx, t)
-	sel := selectors.NewExchangeBackup()
-	sel.Include(sel.MailFolders(selectors.Any(), []string{exchange.DefaultMailFolder}, selectors.PrefixMatch()))
-	scopes := sel.DiscreteScopes(connector.GetUsers())
-
-	for _, scope := range scopes {
-		users := scope.Get(selectors.ExchangeUser)
-		standard := (len(users) / 4) * 3
-		collections, err := connector.createCollections(ctx, scope)
-		require.NoError(t, err)
-		suite.Greater(len(collections), standard)
-	}
 }
 
 func (suite *GraphConnectorIntegrationSuite) TestMailFetch() {
@@ -409,63 +117,9 @@ func (suite *GraphConnectorIntegrationSuite) TestMailFetch() {
 	}
 }
 
-///------------------------------------------------------------
+//-------------------------------------------------------------
 // Exchange Functions
-//-------------------------------------------------------
-
-func (suite *GraphConnectorIntegrationSuite) TestEmptyCollections() {
-	dest := tester.DefaultTestRestoreDestination()
-	table := []struct {
-		name string
-		col  []data.Collection
-		sel  selectors.Selector
-	}{
-		{
-			name: "ExchangeNil",
-			col:  nil,
-			sel: selectors.Selector{
-				Service: selectors.ServiceExchange,
-			},
-		},
-		{
-			name: "ExchangeEmpty",
-			col:  []data.Collection{},
-			sel: selectors.Selector{
-				Service: selectors.ServiceExchange,
-			},
-		},
-		{
-			name: "OneDriveNil",
-			col:  nil,
-			sel: selectors.Selector{
-				Service: selectors.ServiceOneDrive,
-			},
-		},
-		{
-			name: "OneDriveEmpty",
-			col:  []data.Collection{},
-			sel: selectors.Selector{
-				Service: selectors.ServiceOneDrive,
-			},
-		},
-	}
-
-	for _, test := range table {
-		suite.T().Run(test.name, func(t *testing.T) {
-			ctx, flush := tester.NewContext()
-			defer flush()
-
-			deets, err := suite.connector.RestoreDataCollections(ctx, test.sel, dest, test.col)
-			require.NoError(t, err)
-			assert.NotNil(t, deets)
-
-			stats := suite.connector.AwaitStatus()
-			assert.Zero(t, stats.ObjectCount)
-			assert.Zero(t, stats.FolderCount)
-			assert.Zero(t, stats.Successful)
-		})
-	}
-}
+//-------------------------------------------------------------
 
 func runRestoreBackupTest(
 	t *testing.T,
