+++ src/internal/tester/integration_runners.go
@@ -9,21 +9,23 @@ import (
 )
 
 const (
-	CorsoLoadTests                   = "CORSO_LOAD_TESTS"
-	CorsoCITests                     = "CORSO_CI_TESTS"
-	CorsoCLIBackupTests              = "CORSO_COMMAND_LINE_BACKUP_TESTS"
-	CorsoCLIConfigTests              = "CORSO_COMMAND_LINE_CONFIG_TESTS"
-	CorsoCLIRepoTests                = "CORSO_COMMAND_LINE_REPO_TESTS"
-	CorsoCLIRestoreTests             = "CORSO_COMMAND_LINE_RESTORE_TESTS"
-	CorsoCLITests                    = "CORSO_COMMAND_LINE_TESTS"
-	CorsoGraphConnectorTests         = "CORSO_GRAPH_CONNECTOR_TESTS"
-	CorsoGraphConnectorExchangeTests = "CORSO_GRAPH_CONNECTOR_EXCHANGE_TESTS"
-	CorsoGraphConnectorOneDriveTests = "CORSO_GRAPH_CONNECTOR_ONE_DRIVE_TESTS"
-	CorsoKopiaWrapperTests           = "CORSO_KOPIA_WRAPPER_TESTS"
-	CorsoModelStoreTests             = "CORSO_MODEL_STORE_TESTS"
-	CorsoOneDriveTests               = "CORSO_ONE_DRIVE_TESTS"
-	CorsoOperationTests              = "CORSO_OPERATION_TESTS"
-	CorsoRepositoryTests             = "CORSO_REPOSITORY_TESTS"
+	CorsoLoadTests                      = "CORSO_LOAD_TESTS"
+	CorsoCITests                        = "CORSO_CI_TESTS"
+	CorsoCLIBackupTests                 = "CORSO_COMMAND_LINE_BACKUP_TESTS"
+	CorsoCLIConfigTests                 = "CORSO_COMMAND_LINE_CONFIG_TESTS"
+	CorsoCLIRepoTests                   = "CORSO_COMMAND_LINE_REPO_TESTS"
+	CorsoCLIRestoreTests                = "CORSO_COMMAND_LINE_RESTORE_TESTS"
+	CorsoCLITests                       = "CORSO_COMMAND_LINE_TESTS"
+	CorsoConnectorCreateCollectionTests = "CORSO_CONNECTOR_CREATE_COLLECTION_TESTS"
+	CorsoConnectorDataCollectionTests   = "CORSO_CONNECTOR_DATA_COLLECTION_TESTS"
+	CorsoGraphConnectorTests            = "CORSO_GRAPH_CONNECTOR_TESTS"
+	CorsoGraphConnectorExchangeTests    = "CORSO_GRAPH_CONNECTOR_EXCHANGE_TESTS"
+	CorsoGraphConnectorOneDriveTests    = "CORSO_GRAPH_CONNECTOR_ONE_DRIVE_TESTS"
+	CorsoKopiaWrapperTests              = "CORSO_KOPIA_WRAPPER_TESTS"
+	CorsoModelStoreTests                = "CORSO_MODEL_STORE_TESTS"
+	CorsoOneDriveTests                  = "CORSO_ONE_DRIVE_TESTS"
+	CorsoOperationTests                 = "CORSO_OPERATION_TESTS"
+	CorsoRepositoryTests                = "CORSO_REPOSITORY_TESTS"
 )
 
 // File needs to be a single message .json
